requirements.txt
PySide6==6.9.2
pytesseract==0.3.13
Pillow==11.3.0
mss==10.1.0
keyboard==0.13.5
rapidfuzz==3.14.1
pytest==8.4.2
opencv-python==4.12.0.88
